Handle int32 and int64 size results from Koji over XMLRPC (#61)
Up until now we were always getting an int32 back from Koji
representing an RPM size.  But now we might get an int32 or an int64.
If you try to decode an int64 as an int32, xmlrpc-c will crash.  I
have changed the librpminspect size type to long long int and store
the size in that.  I decode based on the value type.  If for some
reason we get something that is not an int32 or an int64, I store -47
in the size so it will act as a reminder that decoding failed and
there is possibly a new type lurking in there.

+++ lib/types.h
@@ -321,7 +321,7 @@ typedef struct _koji_rpmlist_entry_t {
     char *version;
     char *release;
     int epoch;
-    int size;
+    long long int size;
     TAILQ_ENTRY(_koji_rpmlist_entry_t) items;
 } koji_rpmlist_entry_t;
